refactor(notes): split detail.rs into dates, reminders and audio_section modules (#21)
