+++ src/ui/notes/detail.rs
@@ -1,97 +1,25 @@
 use crate::db::Database;
 use crate::models::{
     generate_auto_title, is_auto_title, Attachment, NewAttachment, NewTextNote,
-    NoteAudio, NoteReminder, ReminderIntent, UpdateNote,
+    UpdateNote,
 };
-use crate::services::audio::{self, RecordingState};
+use crate::services::audio::RecordingState;
 use crate::services::embed::{embed_attachment, embed_note};
-use crate::services::i18n::{month_abbr, reminder_due_label, t, t_args};
-use crate::services::transcription::SonioxClient;
+use crate::services::i18n::{t, t_args};
 use crate::ui::folder_picker::FolderPicker;
-use crate::ui::icons::{IconArrowUpRight, IconBell, IconTrash, IconX};
 use crate::ui::notes::attachments::AttachmentSection;
-use crate::ui::notes::audio_player::AudioPlayer;
+use crate::ui::notes::audio_section::{AudioJobBanner, AudioSection};
+use crate::ui::notes::dates::{format_absolute_short, format_relative_date};
 use crate::ui::notes::menu::{
     import_audio_file, import_file_content, NoteMenu,
 };
+use crate::ui::notes::reminders::{ActiveReminders, ReminderSuggestions};
 use crate::ui::notes::tags::TagsSection;
 use crate::ui::recording::RecordingBar;
-use crate::ui::transcription_manager::{JobStatus, TranscriptionManager};
+use crate::ui::transcription_manager::TranscriptionManager;
 use crate::ui::{AppState, View};
-use chrono::{Datelike, Local, NaiveDateTime, NaiveTime, Timelike, Utc};
 use dioxus::prelude::*;
 use std::sync::Arc;
-use std::time::Duration;
-
-fn format_relative_date(iso: &str, lang: &str) -> String {
-    let parsed = NaiveDateTime::parse_from_str(
-        &iso.replace('T', " ").replace('Z', ""),
-        "%Y-%m-%d %H:%M:%S%.f",
-    );
-    let dt = match parsed {
-        Ok(d) => d,
-        Err(_) => return iso.to_string(),
-    };
-    let now = Utc::now().naive_utc();
-    let diff = now.signed_duration_since(dt);
-    let secs = diff.num_seconds();
-
-    if secs < 60 {
-        return t(lang, "date-just-now");
-    }
-    if secs < 3600 {
-        let mins = (secs / 60).to_string();
-        return t_args(lang, "date-mins-ago", &[("mins", &mins)]);
-    }
-    if secs < 86400 {
-        let hours = (secs / 3600).to_string();
-        return t_args(lang, "date-hours-ago", &[("hours", &hours)]);
-    }
-
-    let today = now.date();
-    let note_date = dt.date();
-    if today.pred_opt() == Some(note_date) {
-        let time = if lang == "fr" {
-            dt.format("%H:%M").to_string()
-        } else {
-            dt.format("%-I:%M %p").to_string()
-        };
-        return t_args(lang, "date-yesterday", &[("time", &time)]);
-    }
-
-    let m = month_abbr(lang, note_date.month());
-    let d = note_date.day();
-
-    if note_date.year() == today.year() {
-        format!("{d} {m}")
-    } else {
-        format!("{d} {m} {}", note_date.year())
-    }
-}
-
-fn format_absolute_short(iso: &str, lang: &str) -> String {
-    let parsed = NaiveDateTime::parse_from_str(
-        &iso.replace('T', " ").replace('Z', ""),
-        "%Y-%m-%d %H:%M:%S%.f",
-    );
-    let dt = match parsed {
-        Ok(d) => d,
-        Err(_) => return iso.to_string(),
-    };
-    let d = dt.date();
-    let now = Utc::now().naive_utc().date();
-    let m = month_abbr(lang, d.month());
-    if d.year() == now.year() {
-        let time = if lang == "fr" {
-            dt.format("%H:%M").to_string()
-        } else {
-            dt.format("%-I:%M %p").to_string()
-        };
-        format!("{} {}, {}", d.day(), m, time)
-    } else {
-        format!("{} {} {}", d.day(), m, d.year())
-    }
-}
 
 #[component]
 pub fn NoteDetail() -> Element {
@@ -162,13 +90,6 @@ pub fn NoteDetail() -> Element {
     let mut import_in_progress = use_signal(|| false);
     let mut pending_audio: Signal<Option<(String, f64)>> = use_signal(|| None);
     let mut audios_version = use_signal(|| 0u32);
-    let mut detected_reminders: Signal<Vec<ReminderIntent>> =
-        use_signal(Vec::new);
-    let mut reminders_checked = use_signal(String::new);
-    let mut detecting_reminders = use_signal(|| false);
-    let mut reminder_feedback: Signal<Option<String>> = use_signal(|| None);
-    let mut confirm_delete_reminder: Signal<Option<String>> =
-        use_signal(|| None);
 
     use_effect(move || {
         if !deleted() {
@@ -222,19 +143,6 @@ pub fn NoteDetail() -> Element {
         }
     });
 
-    let audios: Vec<NoteAudio> = {
-        let id = local_note_id();
-        let _ = audios_version();
-        let _ = (app.notes_version)();
-        if id.is_empty() {
-            Vec::new()
-        } else {
-            db().list_audios(&id).unwrap_or_default()
-        }
-    };
-    let mut audios_expanded = use_signal(|| false);
-    let mut transcribing_audio_id: Signal<Option<String>> = use_signal(|| None);
-
     let attachments_version = (app.attachments_version)();
     let attachments: Vec<Attachment> = {
         let id = local_note_id();
@@ -243,19 +151,6 @@ pub fn NoteDetail() -> Element {
             Vec::new()
         } else {
             db().list_attachments_for_note(&id).unwrap_or_default()
-        }
-    };
-
-    let detail_reminders: Vec<NoteReminder> = {
-        let id = local_note_id();
-        let _ = (app.notes_version)();
-        if id.is_empty() {
-            Vec::new()
-        } else {
-            db().reminders_for_note(&id)
-                .into_iter()
-                .filter(|r| r.state == "active")
-                .collect()
         }
     };
 
@@ -387,76 +282,6 @@ pub fn NoteDetail() -> Element {
             }
             generating_title.set(false);
             title_gen_done.set(true);
-        });
-    });
-
-    let reminder_initial_content = initial_content.clone();
-    use_effect(move || {
-        let c = content();
-        if c.trim().chars().count() < 8 {
-            if !detected_reminders.peek().is_empty() {
-                detected_reminders.set(Vec::new());
-            }
-            return;
-        }
-        if *reminders_checked.peek() == c {
-            return;
-        }
-        let nid = local_note_id();
-        if c == reminder_initial_content
-            && !nid.is_empty()
-            && db().note_has_active_reminder(&nid)
-        {
-            reminders_checked.set(c);
-            return;
-        }
-        spawn(async move {
-            futures_timer::Delay::new(Duration::from_millis(1200)).await;
-            if *content.peek() != c {
-                return;
-            }
-            if *reminders_checked.peek() == c || *detecting_reminders.peek() {
-                return;
-            }
-            detecting_reminders.set(true);
-            match crate::services::llm::LlmClient::from_db(&db()) {
-                Ok(ai) => {
-                    if let Ok(intents) =
-                        ai.extract_reminders(&c, Local::now()).await
-                    {
-                        let normalized: Vec<ReminderIntent> = intents
-                            .into_iter()
-                            .map(|mut it| {
-                                if let Some(d) = it.resolved_date() {
-                                    it.date =
-                                        Some(d.format("%Y-%m-%d").to_string());
-                                }
-                                let tt = it.resolved_time();
-                                it.time = Some(format!(
-                                    "{:02}:{:02}",
-                                    tt.hour(),
-                                    tt.minute()
-                                ));
-                                it.time_end =
-                                    it.resolved_time_end().map(|end| {
-                                        format!(
-                                            "{:02}:{:02}",
-                                            end.hour(),
-                                            end.minute()
-                                        )
-                                    });
-                                it
-                            })
-                            .collect();
-                        reminders_checked.set(c.clone());
-                        detected_reminders.set(normalized);
-                    }
-                }
-                Err(_) => {
-                    reminders_checked.set(c.clone());
-                }
-            }
-            detecting_reminders.set(false);
         });
     });
 
@@ -659,18 +484,6 @@ pub fn NoteDetail() -> Element {
         app.transcription_done_badge.set(0);
     });
 
-    let audio_job_status: Option<JobStatus> = {
-        let jobs = (app.transcription_jobs)();
-        let nid = local_note_id();
-        if nid.is_empty() {
-            None
-        } else {
-            jobs.get(&nid)
-                .and_then(|q| q.front())
-                .map(|j| j.status.clone())
-        }
-    };
-
     let recording_state = (app.recording_state)();
     let is_transcribing = recording_state == RecordingState::Transcribing;
 
@@ -688,8 +501,6 @@ pub fn NoteDetail() -> Element {
     let title_placeholder = t(&lang, "note-title-placeholder");
     let content_placeholder = t(&lang, "note-content-placeholder");
     let transcribing_placeholder = t(&lang, "note-transcribing-placeholder");
-    let transcribing_short = t(&lang, "note-transcribing-short");
-    let transcribe_action = t(&lang, "note-transcribe-action");
     let created_on_text = if created_date.is_empty() {
         String::new()
     } else {
@@ -758,82 +569,7 @@ pub fn NoteDetail() -> Element {
                         }
                     }
                 }
-                if !detail_reminders.is_empty() {
-                    div { class: "mt-2 px-1 flex flex-col gap-1.5",
-                        for r in detail_reminders.iter() {
-                            {
-                                let due = reminder_due_label(&lang, r);
-                                let del_label = t(&lang, "reminder-delete");
-                                let cancel_label = t(&lang, "reminder-cancel");
-                                let row_id = r.id.clone();
-                                let event_id_open = r.reminder_id.clone();
-                                let event_id_del = r.reminder_id.clone();
-                                let confirming = confirm_delete_reminder() == Some(row_id.clone());
-                                rsx! {
-                                    div { class: "flex items-center gap-1.5",
-                                        button {
-                                            class: "flex-1 flex items-center gap-2 min-h-[44px] px-3 py-2 rounded-xl bg-ios-orange/10 text-ios-orange-dark text-xs font-medium active:bg-ios-orange/25",
-                                            onclick: move |_| {
-                                                #[cfg(target_os = "ios")]
-                                                {
-                                                    let event_id_open = event_id_open.clone();
-                                                    spawn(async move {
-                                                        crate::platform::ios::reminders::present_event(event_id_open).await;
-                                                    });
-                                                }
-                                                #[cfg(not(target_os = "ios"))]
-                                                let _ = &event_id_open;
-                                            },
-                                            IconBell { size: 15 }
-                                            span { class: "flex-1 text-left", "{due}" }
-                                            IconArrowUpRight { size: 14 }
-                                        }
-                                        if confirming {
-                                            button {
-                                                class: "shrink-0 min-h-[44px] px-3 rounded-xl text-xs font-medium text-white bg-ios-red active:opacity-70",
-                                                onclick: {
-                                                    let row_id = row_id.clone();
-                                                    let event_id_del = event_id_del.clone();
-                                                    move |_| {
-                                                        confirm_delete_reminder.set(None);
-                                                        let _ = db().delete_note_reminder(&row_id);
-                                                        app.notes_version.set((app.notes_version)() + 1);
-                                                        #[cfg(target_os = "ios")]
-                                                        {
-                                                            let event_id_del = event_id_del.clone();
-                                                            spawn(async move {
-                                                                if let Err(e) = crate::platform::ios::reminders::remove_event(event_id_del).await {
-                                                                    eprintln!("[reminder] remove_event failed: {e}");
-                                                                }
-                                                            });
-                                                        }
-                                                        #[cfg(not(target_os = "ios"))]
-                                                        let _ = &event_id_del;
-                                                    }
-                                                },
-                                                "{del_label}"
-                                            }
-                                            button {
-                                                class: "shrink-0 min-h-[44px] px-2 text-xs font-medium text-stone-500 active:opacity-70",
-                                                onclick: move |_| confirm_delete_reminder.set(None),
-                                                "{cancel_label}"
-                                            }
-                                        } else {
-                                            button {
-                                                class: "shrink-0 min-h-[44px] p-2 text-stone-400 active:text-ios-red",
-                                                onclick: {
-                                                    let row_id = row_id.clone();
-                                                    move |_| confirm_delete_reminder.set(Some(row_id.clone()))
-                                                },
-                                                IconTrash { size: 16 }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
+                ActiveReminders { local_note_id }
             }
             div { class: "border-t border-stone-100 pt-3 pb-2",
                 TagsSection { tags, tag_input, tagging, content }
@@ -848,269 +584,14 @@ pub fn NoteDetail() -> Element {
                 value: "{content}",
                 oninput: move |evt| content.set(evt.value()),
             }
-            {
-                let detected = detected_reminders();
-                let pending_idx: Vec<usize> = detected
-                    .iter()
-                    .enumerate()
-                    .filter(|(_, i)| {
-                        let nid = local_note_id();
-                        nid.is_empty()
-                            || !db().reminder_exists_by_intent_hash(
-                                &nid,
-                                &i.intent_hash(),
-                            )
-                    })
-                    .map(|(idx, _)| idx)
-                    .collect();
-                let title_label = t(&lang, "reminder-detected-title");
-                let confirm_label = t(&lang, "reminder-confirm");
-                let range_add_label = t(&lang, "reminder-range-add");
-                let range_to_label = t(&lang, "reminder-range-to");
-                let lang_badge = lang.clone();
-                rsx! {
-                    if !pending_idx.is_empty() {
-                            div { class: "mt-3 p-3 bg-ios-orange/10 border border-ios-orange/30 rounded-xl",
-                            div { class: "flex items-center justify-between mb-2",
-                                span { class: "text-xs font-semibold text-ios-orange-dark", "{title_label}" }
-                                button {
-                                    class: "text-stone-400 active:opacity-70",
-                                    onclick: move |_| {
-                                        reminders_checked.set(content.peek().clone());
-                                        detected_reminders.set(Vec::new());
-                                    },
-                                    IconX { size: 14 }
-                                }
-                            }
-                            div { class: "space-y-2",
-                                for idx in pending_idx {
-                                    {
-                                        let intent = detected[idx].clone();
-                                        let action = intent.action.clone();
-                                        let date_val = intent.date.clone().unwrap_or_default();
-                                        let time_val = intent.time.clone().unwrap_or_default();
-                                        let end_val = intent.time_end.clone().unwrap_or_default();
-                                        let has_end = intent.time_end.is_some();
-                                        let default_end = {
-                                            use chrono::Duration as ChronoDuration;
-                                            NaiveTime::parse_from_str(&time_val, "%H:%M")
-                                                .ok()
-                                                .map(|tm| tm + ChronoDuration::hours(1))
-                                                .map(|tm| format!("{:02}:{:02}", tm.hour(), tm.minute()))
-                                                .unwrap_or_else(|| "10:00".to_string())
-                                        };
-                                        let is_rec = intent.recurrence.is_some();
-                                        let lang_cb = lang_badge.clone();
-                                        let confirm = confirm_label.clone();
-                                        let range_add = range_add_label.clone();
-                                        let range_to = range_to_label.clone();
-                                        rsx! {
-                                            div { class: "bg-white/40 rounded-lg p-2.5",
-                                                p { class: "text-sm text-stone-800 mb-1.5", "{action}" }
-                                                div { class: "flex items-center gap-2 flex-wrap",
-                                                    input {
-                                                        r#type: "date",
-                                                        class: "text-xs text-stone-700 bg-white border border-stone-200 rounded-md px-2 py-1 outline-none",
-                                                        value: "{date_val}",
-                                                        oninput: move |e| {
-                                                            detected_reminders.write()[idx].date = Some(e.value());
-                                                        },
-                                                    }
-                                                    input {
-                                                        r#type: "time",
-                                                        class: "text-xs text-stone-700 bg-white border border-stone-200 rounded-md px-2 py-1 outline-none",
-                                                        value: "{time_val}",
-                                                        oninput: move |e| {
-                                                            detected_reminders.write()[idx].time = Some(e.value());
-                                                        },
-                                                    }
-                                                    if has_end {
-                                                        span { class: "text-xs text-stone-500", "{range_to}" }
-                                                        input {
-                                                            r#type: "time",
-                                                            class: "text-xs text-stone-700 bg-white border border-stone-200 rounded-md px-2 py-1 outline-none",
-                                                            value: "{end_val}",
-                                                            oninput: move |e| {
-                                                                detected_reminders.write()[idx].time_end = Some(e.value());
-                                                            },
-                                                        }
-                                                        button {
-                                                            class: "text-stone-400 active:opacity-70",
-                                                            onclick: move |_| {
-                                                                detected_reminders.write()[idx].time_end = None;
-                                                            },
-                                                            IconX { size: 12 }
-                                                        }
-                                                    } else {
-                                                        button {
-                                                            class: "text-xs font-medium text-ios-orange-dark active:opacity-70",
-                                                            onclick: move |_| {
-                                                                detected_reminders.write()[idx].time_end = Some(default_end.clone());
-                                                            },
-                                                            "{range_add}"
-                                                        }
-                                                    }
-                                                    if is_rec {
-                                                        span { class: "text-xs text-ios-orange-dark", "↻" }
-                                                    }
-                                                    button {
-                                                    class: "shrink-0 self-center text-xs font-medium text-white bg-ios-orange-dark px-3 py-1.5 rounded-full active:opacity-70",
-                                                    onclick: move |_| {
-                                                        let intent = detected_reminders.peek()[idx].clone();
-                                                        let lang2 = lang_cb.clone();
-                                                        let database = db();
-                                                        spawn(async move {
-                                                            reminder_feedback.set(Some(t(&lang2, "reminder-creating")));
-                                                            let nid = {
-                                                                let cur = local_note_id();
-                                                                if cur.is_empty() {
-                                                                    let tt = title();
-                                                                    let new = NewTextNote {
-                                                                        title: if tt.is_empty() { None } else { Some(tt) },
-                                                                        content: content(),
-                                                                        tags: tags(),
-                                                                    };
-                                                                    match database.create_text_note(&new) {
-                                                                        Ok(created) => {
-                                                                            if let Some(ref fid) = (app.detail_folder_id)() {
-                                                                                let _ = database.add_note_to_folder(&created.id, fid);
-                                                                            }
-                                                                            local_note_id.set(created.id.clone());
-                                                                            app.current_note_id.set(Some(created.id.clone()));
-                                                                            app.notes_version.set((app.notes_version)() + 1);
-                                                                            created.id
-                                                                        }
-                                                                        Err(e) => {
-                                                                            reminder_feedback.set(Some(t_args(&lang2, "reminder-failed", &[("error", &e)])));
-                                                                            return;
-                                                                        }
-                                                                    }
-                                                                } else {
-                                                                    cur
-                                                                }
-                                                            };
-                                                            let res = crate::services::reminders::schedule(database, nid, intent.clone()).await;
-                                                            use crate::services::reminders::ScheduleResult;
-                                                            match res {
-                                                                ScheduleResult::Created => {
-                                                                    app.notes_version.set((app.notes_version)() + 1);
-                                                                    reminder_feedback.set(Some(t(&lang2, "reminder-created")));
-                                                                }
-                                                                ScheduleResult::Duplicate => reminder_feedback.set(Some(t(&lang2, "reminder-duplicate"))),
-                                                                ScheduleResult::AccessDenied => reminder_feedback.set(Some(t(&lang2, "reminder-denied"))),
-                                                                ScheduleResult::Unsupported => reminder_feedback.set(Some("iOS only".to_string())),
-                                                                ScheduleResult::Failed(e) => reminder_feedback.set(Some(t_args(&lang2, "reminder-failed", &[("error", &e)]))),
-                                                            }
-                                                        });
-                                                    },
-                                                    "{confirm}"
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        }
-                    }
-                }
+            ReminderSuggestions {
+                local_note_id,
+                title,
+                content,
+                tags,
+                initial_content: initial_content.clone(),
             }
-            if let Some(msg) = reminder_feedback() {
-                div {
-                    class: "mt-2 px-3 py-2 bg-ios-orange/10 rounded-lg text-xs text-stone-600",
-                    "{msg}"
-                }
-            }
-            if !audios.is_empty() {
-                {
-                    let count = audios.len().to_string();
-                    let audio_label = t_args(&lang, "note-audios-label", &[("count", &count)]);
-                    rsx! {
-                        div { class: "mt-3",
-                            button {
-                                class: "flex items-center gap-2 w-full py-2 text-left active:opacity-70 transition-opacity duration-150",
-                                onclick: move |_| audios_expanded.set(!audios_expanded()),
-                                span { class: "text-xs font-medium text-stone-500", "{audio_label}" }
-                                span {
-                                    class: if audios_expanded() {
-                                        "inline-block w-1.5 h-1.5 border-r-[1.5px] border-b-[1.5px] border-stone-400 transition-transform duration-150 -rotate-[135deg]"
-                                    } else {
-                                        "inline-block w-1.5 h-1.5 border-r-[1.5px] border-b-[1.5px] border-stone-400 transition-transform duration-150 rotate-45"
-                                    },
-                                }
-                            }
-                            if audios_expanded() {
-                                div { class: "space-y-2 pt-1",
-                                    for audio in audios.iter() {
-                                        {
-                                            let audio_id = audio.id.clone();
-                                            let audio_id_tr = audio.id.clone();
-                                            let file_path = audio.file_path.clone();
-                                            let file_path_tr = audio.file_path.clone();
-                                            let resolved = audio::resolve_audio_path(&file_path);
-                                            let date = format_relative_date(&audio.created_at, &lang);
-                                            let transcription = audio.transcription.clone();
-                                            let is_transcribing_this = transcribing_audio_id() == Some(audio.id.clone());
-                                            rsx! {
-                                                div { class: "space-y-1",
-                                                    p { class: "text-[10px] text-stone-400 px-1", "{date}" }
-                                                    AudioPlayer {
-                                                        audio_path: resolved,
-                                                        duration_secs: audio.duration_secs,
-                                                        on_delete: move |_| {
-                                                            let fp = file_path.clone();
-                                                            let _ = std::fs::remove_file(audio::resolve_audio_path(&fp));
-                                                            let _ = db().delete_audio(&audio_id);
-                                                            audios_version.set(audios_version() + 1);
-                                                            app.notes_version.set((app.notes_version)() + 1);
-                                                        },
-                                                    }
-                                                    if let Some(ref text) = transcription {
-                                                        p { class: "text-xs text-stone-600 italic px-1 pb-1", "{text}" }
-                                                    } else if is_transcribing_this {
-                                                        p {
-                                                            class: "text-xs text-stone-400 px-1 pb-1",
-                                                            style: "animation: pulseSoft 1.5s ease-in-out infinite;",
-                                                            "{transcribing_short}"
-                                                        }
-                                                    } else {
-                                                        button {
-                                                            class: "text-xs text-ios-orange-dark px-1 pb-1 active:opacity-70",
-                                                            onclick: move |_| {
-                                                                let aid = audio_id_tr.clone();
-                                                                let fp = file_path_tr.clone();
-                                                                let path = audio::resolve_audio_path(&fp);
-                                                                transcribing_audio_id.set(Some(aid.clone()));
-                                                                spawn(async move {
-                                                                    let result = async {
-                                                                        let client = SonioxClient::from_db(&db())?;
-                                                                        client.transcribe(std::path::Path::new(&path), None).await
-                                                                    }.await;
-                                                                    match result {
-                                                                        Ok(text) => {
-                                                                            let _ = db().set_audio_transcription(&aid, &text);
-                                                                        }
-                                                                        Err(e) => {
-                                                                            eprintln!("[transcribe] error: {e}");
-                                                                        }
-                                                                    }
-                                                                    transcribing_audio_id.set(None);
-                                                                    audios_version.set(audios_version() + 1);
-                                                                });
-                                                            },
-                                                            "{transcribe_action}"
-                                                        }
-                                                    }
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
+            AudioSection { local_note_id, audios_version }
             AttachmentSection { attachments, confirm_delete_att }
             if let Some(ref status) = import_status() {
                 div { class: if is_importing {
@@ -1124,54 +605,7 @@ pub fn NoteDetail() -> Element {
                     span { class: "text-xs text-stone-600", "{status}" }
                 }
             }
-            if let Some(ref js) = audio_job_status {
-                match js {
-                    JobStatus::Failed(reason) => {
-                        let failed_label = t(&lang, "audio-import-failed");
-                        let retry_label = t(&lang, "audio-import-retry");
-                        let msg = format!("{failed_label}: {reason}");
-                        let retry_manager = manager.clone();
-                        let dismiss_manager = manager.clone();
-                        rsx! {
-                            div { class: "flex items-center gap-2 px-3 py-2 mt-2 bg-ios-red/10 rounded-lg",
-                                span { class: "text-xs text-stone-600 flex-1", "{msg}" }
-                                button {
-                                    class: "text-xs text-ios-orange-dark font-medium active:opacity-70",
-                                    onclick: move |_| {
-                                        let nid = local_note_id();
-                                        retry_manager.retry(&nid);
-                                    },
-                                    "{retry_label}"
-                                }
-                                button {
-                                    class: "text-stone-400 active:opacity-70",
-                                    onclick: move |_| {
-                                        let nid = local_note_id();
-                                        dismiss_manager.dismiss(&nid);
-                                    },
-                                    IconX { size: 14 }
-                                }
-                            }
-                        }
-                    }
-                    JobStatus::Done(_) => rsx! {},
-                    other => {
-                        let transcribing_label = t(&lang, "audio-transcribing");
-                        let label = match other {
-                            JobStatus::Polling { elapsed_s } => {
-                                format!("{transcribing_label} · {}:{:02}", elapsed_s / 60, elapsed_s % 60)
-                            }
-                            _ => transcribing_label,
-                        };
-                        rsx! {
-                            div { class: "flex items-center gap-2 px-3 py-2 mt-2 bg-ios-orange/10 rounded-lg",
-                                span { class: "inline-block w-3 h-3 border-2 border-ios-orange border-t-transparent rounded-full animate-spin" }
-                                span { class: "text-xs text-stone-600", "{label}" }
-                            }
-                        }
-                    }
-                }
-            }
+            AudioJobBanner { local_note_id }
             if let RecordingState::Error(ref e) = recording_state {
                 {
                     let msg = e.clone();
