+++ src/ui/notes/mod.rs
@@ -1,7 +1,10 @@
 mod attachments;
 mod audio_player;
+mod audio_section;
+mod dates;
 mod detail;
 mod menu;
+mod reminders;
 mod tags;
 
 pub use detail::NoteDetail;
